fix modal background
